BAEL-835 - code rerfactor
