+++ core-java/src/main/java/com/baeldung/outofmemoryerror/OutOfMemoryGCLimitExceed.java
@@ -1,0 +1,19 @@
+package com.baeldung.outofmemoryerror;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Random;
+
+public class OutOfMemoryGCLimitExceed {
+    public static void addRandomDataToMap() {
+        Map<Integer, String> dataMap = new HashMap<>();
+        Random r = new Random();
+        while (true) {
+            dataMap.put(r.nextInt(), String.valueOf(r.nextInt()));
+        }
+    }
+
+    public static void main(String[] args) {
+        OutOfMemoryGCLimitExceed.addRandomDataToMap();
+    }
+}
