--- core-java/src/main/java/com/baeldung/outofmemoryerror/java-outofmemoryerror-demo/pom.xml
@@ -1,44 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <groupId>com.baeldung.outofmemoryerror</groupId>
-  <artifactId>java-outofmemoryerror-demo</artifactId>
-  <version>0.0.1-SNAPSHOT</version>
-  <packaging>jar</packaging>
-
-  <name>java-outofmemoryerror-demo</name>
-  <url>http://maven.apache.org</url>
-
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-  </properties>
-
-  <build>
-   <plugins>
-	   <plugin>
-	        <groupId>org.codehaus.mojo</groupId>
-	        <artifactId>exec-maven-plugin</artifactId>
-	        <version>1.2</version>
-	        <configuration>
-	          <executable>java</executable>
-	          <arguments>
-	            <argument>-Xmx100m</argument>
-	            <argument>-XX:+UseParallelGC</argument>
-	            <argument>-classpath</argument>
-	            <classpath/>
-	            <argument>com.baeldung.outofmemoryerror.java_outofmemoryerror_demo.Test</argument>
-	          </arguments>
-	        </configuration>
-	      </plugin>
-   </plugins>
-  </build>
-  <dependencies>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>3.8.1</version>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-</project>
