--- core-java/src/main/java/com/baeldung/outofmemoryerror/java-outofmemoryerror-demo/src/main/java/com/baeldung/outofmemoryerror/java_outofmemoryerror_demo/Test.java
@@ -1,8 +1,0 @@
-package com.baeldung.outofmemoryerror.java_outofmemoryerror_demo;
-
-public class Test {
-    public static void main( String[] args )
-    {
-        OutOfMemoryGCLimitExceed.addRandomDataToMap();
-    }
-}
